Marge fork

+++ artemis-protocols/artemis-openwire-protocol/src/main/java/org/apache/activemq/artemis/core/protocol/openwire/OpenWireConnection.java
@@ -863,12 +863,14 @@ public class OpenWireConnection extends AbstractRemotingConnection implements Se
          }
 
          List<AMQConsumer> consumersList = amqSession.createConsumer(info, new SlowConsumerDetection());
+
+         this.addConsumerBrokerExchange(info.getConsumerId(), amqSession, consumersList);
+         ss.addConsumer(info);
+
          if (consumersList.size() == 0) {
             return;
          }
 
-         this.addConsumerBrokerExchange(info.getConsumerId(), amqSession, consumersList);
-         ss.addConsumer(info);
          amqSession.start();
 
          if (AdvisorySupport.isAdvisoryTopic(info.getDestination())) {
@@ -1033,17 +1035,19 @@ public class OpenWireConnection extends AbstractRemotingConnection implements Se
 
 
       } else {
-         Bindings bindings = server.getPostOffice().getBindingsForAddress(new SimpleString(dest.getPhysicalName()));
+         Bindings bindings = server.getPostOffice().lookupBindingsForAddress(new SimpleString(dest.getPhysicalName()));
 
-         for (Binding binding : bindings.getBindings()) {
-            Queue b = (Queue) binding.getBindable();
-            if (b.getConsumerCount() > 0) {
-               throw new Exception("Destination still has an active subscription: " + dest.getPhysicalName());
+         if (bindings != null) {
+            for (Binding binding : bindings.getBindings()) {
+               Queue b = (Queue) binding.getBindable();
+               if (b.getConsumerCount() > 0) {
+                  throw new Exception("Destination still has an active subscription: " + dest.getPhysicalName());
+               }
+               if (b.isDurable()) {
+                  throw new Exception("Destination still has durable subscription: " + dest.getPhysicalName());
+               }
+               b.deleteQueue();
             }
-            if (b.isDurable()) {
-               throw new Exception("Destination still has durable subscription: " + dest.getPhysicalName());
-            }
-            b.deleteQueue();
          }
       }
 
@@ -1261,7 +1265,9 @@ public class OpenWireConnection extends AbstractRemotingConnection implements Se
                }
             }
          } else {
-            tx.rollback();
+            if (tx != null) {
+               tx.rollback();
+            }
          }
 
          return null;
@@ -1413,7 +1419,9 @@ public class OpenWireConnection extends AbstractRemotingConnection implements Se
                }
             }
          } else {
-            tx.commit(onePhase);
+            if (tx != null) {
+               tx.commit(onePhase);
+            }
          }
 
          return null;
@@ -1464,7 +1472,7 @@ public class OpenWireConnection extends AbstractRemotingConnection implements Se
             internalSession.resetTX(null);
          }
 
-         return new IntegerResponse(XAResource.XA_RDONLY);
+         return new IntegerResponse(XAResource.XA_OK);
       }
 
       @Override
